Factor directory iterator as a trait so that other extractor can rely on it. and make extension generic so that test (<extension> === 'php') can be overriden if needed.
